Update surface functions.py

+++ utils/surface functions.py
@@ -112,7 +112,7 @@ def n_or_more_neg(D, teacher, rad, student, T, n, lr_1, lr_2, steps,experiment_p
       for i in range(n,T+1):
         P += scipy.special.binom(T,i) * p_correct**i * (1-p_correct)**(T-i)
       
-      data['p'][step/D -1] = P
+      data['p'][int(step/D) -1] = P
 
     step += 1
 
@@ -168,7 +168,7 @@ def all_neg(D, teacher, student, T, lr_1, lr_2, steps, experiment_path):
       data['q'][int(step/D) -1] = Q
       
       P = p_T_correct(Q,R,T)
-      data['p'][step/D -1] = P
+      data['p'][int(step/D) -1] = P
 
     step += 1
